requirements.txt
pdfminer.six
pypasser
pyspark
requests
selenium
